Add basic self-hosted Aurelia app
Creating the shell around the Aurelia app, using a modified verison of the es2016 Aurelia Skeleton application, and having the app be served through browserSync since we aren't sure if everyone will have VS 2015 which we'd want if we were hosting through ASP.NET Core, and we decided we didn't want to host through 4.5.  Obviously that can all change, but it shouldn't have an impact on the development of the actual Aurelia app.